Began implementing folders
Completed UI required for for the most part. Wrote the model for a
Folder after discussion with David on the req props. Began work on the
FolderController. Made a root folder to hold the initial file dump on
page load, will then make ajax calls when a folder is selected and the
table will be reloaded with said selected folders items (files &
folders)